LibWeb: Import wpt-test csp-violations file with some modifications
Our WebSocket implementation is not yet hooked into our fetch api.
This means that no policy violations are emitted when using WebSocket.

There is a good chunk of test that emit a final connect-src violation
to signify the end of all violations. We can simply replace this for a
fetch call.